add goerli test case
